+++ HISTORY.rst
@@ -7,5 +7,4 @@ History
 ++++++++++++++++++
 
 * First release on PyPI.
-* Initial commit.
-* lay the ground for the initial project
+* Implemented the basic functionality for job search with keyword, location, company